refactor(optics): rename utils.py → ops.py; inline cosine scheduler; misc cleanup
- deeplens/optics/utils.py → ops.py: file only contains pure differentiable
  torch ops (interp1d, grid_sample_xy, foc_dist_balanced, diff_float,
  diff_quantize); added module docstring to prevent future drift; dropped
  numpy import (only wave_rgb needed it)
- Removed wave_rgb() from ops.py — callers can compose from WAVE_RED/GREEN/BLUE
  constants already in config.py; removed TestWaveRgb class from test_basics.py
- geolens_pkg/optim.py: replaced `from transformers import
  get_cosine_schedule_with_warmup` with a local implementation using only
  math + torch; dropped transformers dependency
- geolens_pkg/io.py: removed redundant self.post_computation() call from
  read_lens_json() and updated docstring accordingly
- Updated all import sites (hybridlens, diffraclens, psf_compute, diffractive,
  test_utils, test_basics) to import from ops instead of utils
