+++ deeplens/optics/geolens_pkg/psf_compute.py
@@ -38,7 +38,7 @@ from ..config import (
 )
 from ..imgsim import forward_integral
 from ..light import AngularSpectrumMethod
-from ..utils import diff_float
+from ..ops import diff_float
 
 
 class GeoLensPSF:
